requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
eval_type_backport==0.2.2
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
huggingface-hub==0.29.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
propcache==0.3.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
shellingham==1.5.4
tabulate==0.9.0
together==1.4.1
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
streamlit
nltk==3.6.7
numpy==1.21.2
tensorflow==2.6.0
python-dotenv==0.19.0